+++ src/scripts/endScreen.js
@@ -11,15 +11,15 @@ define(function(require) {
 			document.getElementById("end-results-mapper-name").textContent = `Mapped by ${playDetails.mapperName}`;
 			document.getElementById("end-results-accuracy-header").textContent = playDetails.accuracy.toFixed(2) + "%";
 			document.getElementById("end-results-max-combo").textContent = playDetails.maxCombo;
-			document.getElementById("end-results-combo-type").textContent = playDetails.fc;
+			document.getElementById("end-results-combo-type").textContent = playDetails.comboType;
 			document.getElementById("end-results-pp").textContent = Math.round(playDetails.pp);
-			document.getElementById("end-results-great").textContent = playDetails.hitDetails.total300;
-			document.getElementById("end-results-ok").textContent = playDetails.hitDetails.total100;
-			document.getElementById("end-results-meh").textContent = playDetails.hitDetails.total50;
-			document.getElementById("end-results-miss").textContent = playDetails.hitDetails.totalMiss;
-			document.getElementById("end-results-slider-ends").textContent = playDetails.hitDetails.hitSliderEnds + "/" + playDetails.hitDetails.totalSliderEnds;
+			document.getElementById("end-results-great").textContent = playDetails.great;
+			document.getElementById("end-results-ok").textContent = playDetails.ok;
+			document.getElementById("end-results-meh").textContent = playDetails.meh;
+			document.getElementById("end-results-miss").textContent = playDetails.miss;
+			document.getElementById("end-results-slider-ends").textContent = playDetails.sliderBreaks;
 			document.getElementById("end-results-unstable-rate").textContent = Math.round(playDetails.unstableRate);
-			document.getElementById("end-results-combo-breaks").textContent = playDetails.hitDetails.comboBreaks;
+			document.getElementById("end-results-combo-breaks").textContent = playDetails.comboBreaks;
 			document.getElementById("end-results-date-played").textContent = "Played on " + playDetails.datePlayed;
 		}
 	}