+++ src/scripts/hitObjects.js
@@ -108,6 +108,22 @@ define(function(require) {
 				this.x = x;
 				this.y = y;
 			}
+		},
+		MouseReplayEvent: class MouseReplayEvent {
+			constructor(x, y, time) {
+				this.type = "mouse";
+				this.time = time;
+				this.x = x;
+				this.y = y;
+			}
+		},
+		KeyboardReplayEvent: class KeyboardReplayEvent {
+			constructor(key, state, time) {
+				this.type = "keyboard";
+				this.time = time;
+				this.key = key;
+				this.state = state;
+			}
 		}
 	};
 });